+++ spotbugs/src/main/java/edu/umd/cs/findbugs/detect/FormatStringChecker.java
@@ -106,11 +106,13 @@ public class FormatStringChecker extends OpcodeStackDetector {
             XMethod m = getXMethodOperand();
             if ((m == null || m.isVarArgs())
                     && sig.indexOf("Ljava/lang/String;[Ljava/lang/Object;)") >= 0
-                    && ("java/util/Formatter".equals(cl) && "format".equals(nm) || "java/lang/String".equals(cl)
-                            && "format".equals(nm) || "java/io/PrintStream".equals(cl) && "format".equals(nm)
-                            || "java/io/PrintStream".equals(cl) && "printf".equals(nm) || cl.endsWith("Writer")
-                                    && "format".equals(nm) || cl.endsWith("Writer") && "printf".equals(nm)) || cl.endsWith("Logger")
-                                            && nm.endsWith("fmt")) {
+                    && ("java/util/Formatter".equals(cl) && "format".equals(nm)
+                            || "java/lang/String".equals(cl) && "format".equals(nm)
+                            || "java/io/PrintStream".equals(cl) && "format".equals(nm)
+                            || "java/io/PrintStream".equals(cl) && "printf".equals(nm)
+                            || cl.endsWith("Writer") && "format".equals(nm)
+                            || cl.endsWith("Writer") && "printf".equals(nm))
+                    || cl.endsWith("Logger") && nm.endsWith("fmt")) {
 
                 if (formatString.indexOf('\n') >= 0) {
                     bugReporter.reportBug(new BugInstance(this, "VA_FORMAT_STRING_USES_NEWLINE", NORMAL_PRIORITY)
@@ -118,7 +120,24 @@ public class FormatStringChecker extends OpcodeStackDetector {
                             .describe(StringAnnotation.FORMAT_STRING_ROLE).addSourceLine(this));
                 }
             }
+        }
 
+        // Check for calls to java.lang.String.formatted(), for instance: "xyz".formatted("abc");
+        if (seen == Const.INVOKEVIRTUAL) {
+            String cl = getClassConstantOperand();
+            String nm = getNameConstantOperand();
+
+            if ("java/lang/String".equals(cl) && "formatted".equals(nm)) {
+                Object stackItem = stack.getStackItem(1).getConstant();
+                if (stackItem instanceof String) {
+                    String string = (String) stackItem;
+                    if (string.indexOf('\n') >= 0) {
+                        bugReporter.reportBug(new BugInstance(this, "VA_FORMAT_STRING_USES_NEWLINE", NORMAL_PRIORITY)
+                                .addClassAndMethod(this).addCalledMethod(this).addString(string)
+                                .describe(StringAnnotation.FORMAT_STRING_ROLE).addSourceLine(this));
+                    }
+                }
+            }
         }
     }